[Refactor] Remove outdated code & add test spec
