--- src/main/scala/BIDMach/allreduce/GridMessages.scala
@@ -1,9 +1,0 @@
-package BIDMach.allreduce
-
-import akka.actor.ActorRef
-
-
-// worker messages
-final case class GridGroupAddresses(group: GridGroup, addresses: Set[ActorRef])
-final case class HelloFromGroup(text: String)
-final case class GreetGroups()
